assistant-toolkit: co-locate Agent with Chat (break the module import cycle)
